backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
django-utils-six==2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
psycopg==3.1.18
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
smartystreets_python_sdk==4.14.2
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
